requirements.txt
PySimpleGUI~=4.60.4
clipboard~=0.0.4
Pillow~=9.3.0